refactor: stories 파일들을 컴포넌트 가까이에 위치하도록 수정
