+++ src/components/stories/App.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import App from "../App";
+import App from "../../App";
 
 const meta = {
   title: "App",
